rather critical additions
add try catch to walk function.

pass file name/info to the determine function methods so appropriate
errors can be generated

stop creating fresh instance var of currentfunctionboundaries as this
was causing boundaries to spill over from previous files when a new
file was loaded. finally this caused doc comments on the first function
to not be picked up as the previous marker would be logically invalid